refactor: Cars 클래스 기능별로 클래스 분리
우승자 출력 메서드를 별도의 클래스에 분리
방어적 복사를 사용
